requirements.txt
flask
flask-sqlalchemy
postgres
python-dotenv
flask_cors